Refactor ApplicantReferenceAcceptanceEmail
